chore: move userSettings into separte slice
