+++ src/components/GearOptimizer.jsx
@@ -10,7 +10,8 @@ import presetExtras from '../assets/presetdata/preset-extras.yaml';
 import presetInfusions from '../assets/presetdata/preset-infusions.yaml';
 import presetTraits from '../assets/presetdata/preset-traits.yaml';
 import templates from '../assets/presetdata/templates.yaml';
-import { getExpertMode, getProfession } from '../state/slices/controlsSlice';
+import { getProfession } from '../state/slices/controlsSlice';
+import { getExpertMode } from '../state/slices/userSettings';
 import NavBar from './nav/NavBar';
 import BossSection from './sections/boss/BossSection';
 import BuffsSection from './sections/buffs/BuffsSection';
