+++ src/components/sections/infusions/Infusions.jsx
@@ -5,7 +5,7 @@ import React from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
 import { makeStyles } from 'tss-react/mui';
-import { getGameMode } from '../../../state/slices/controlsSlice';
+import { getGameMode } from '../../../state/slices/userSettings';
 import {
   changeAR,
   changeInfusion,
